requirements.txt
flas
requests
pandas
numpy
scikit-learn
matplotlib
seaborn
sqlalchemy
psycopg2-binary
gunicorn
transformers
torch
torchvision
pillow
opencv-python
matplotlib